Fixing Issue 2306 addressing a number of issues with • showing lowest terms • fixing bug where only last exercise type of 3 was showing • adding different variable letter sets
Thanks to joshua.olson@gmail.com for fixing/cleaning the file!

The only change I made to your suggested revision was altering a few calls to formatCoefficient for a factored X term and making them formatConstant because, with the x term factored out, the result is a constant and would need an explicit '1' when it was equal to one (as opposed to an implicit and undisplayed 1).

git-svn-id: svn://localhost/trunk@1060 845b83de-3c92-410c-a77f-d821780b4e6f